issue #124 refactor: change MainContainer to LayoutMain and move directory styled-component가 jsx파일 내에 같이 잘못된 명명법으로 선언되어 있어 폴더 생성 후 함꼐 이동
